requirements.txt
Flask==2.2.5
flask_cors==5.0.0
openai==1.55.0
python_dotenv==1.0.1
mongoengine==0.29.1
pymongo=4.10.1